+++ metametameta/from_importlib.py
@@ -2,6 +2,8 @@
 Generate an __about__.py file from package metadata using importlib.metadata.
 """
 
+from __future__ import annotations
+
 import importlib.metadata as md
 import logging
 from typing import Any, cast
@@ -16,7 +18,8 @@ def get_package_metadata(package_name: str) -> dict[str, Any]:
     """Get package metadata using importlib.metadata."""
     try:
         pkg_metadata: md.PackageMetadata = md.metadata(package_name)
-        return {key: value for key, value in cast(dict[str, Any], pkg_metadata).items()}
+        # dict for 3.8 suport
+        return {key: value for key, value in cast(dict, pkg_metadata).items()}
     except md.PackageNotFoundError:
         print(f"Package '{package_name}' not found.")
         return {}