+++ metametameta/known.py
@@ -2,6 +2,8 @@
 Known metadata fields, as opposed to ad hoc ones people make up.
 """
 
+from __future__ import annotations
+
 # https://web.archive.org/web/20111010053227/http://jaynes.colorado.edu/PythonGuidelines.html#module_formatting
 meta = [
     "name",  # title