+++ tests/test_human/test_filesystem.py
@@ -1,4 +1,6 @@
-import os
+from __future__ import annotations
+
+from pathlib import Path
 
 from metametameta.filesystem import write_to_file
 
@@ -10,15 +12,32 @@ def test_write_to_file(tmp_path):
     output_file_name = "__about__.py"
 
     # Execution: Call the function
-    write_to_file(str(directory), about_content, output_file_name)
+    the_result = write_to_file(str(directory), about_content, output_file_name)
+    assert Path(the_result).exists()
+    assert Path(the_result).is_file()
 
-    # Expected output file path
-    expected_file_path = directory / output_file_name
-
-    # Assertion: Check if the file is created at the correct location
-    assert os.path.isfile(expected_file_path), "File was not created at the expected location"
-
-    # Assertion: Check if the content of the file is as expected
-    with open(expected_file_path, encoding="utf-8") as file:
+    with open(the_result, encoding="utf-8") as file:
         content = file.read()
     assert content == about_content, "Content of the file does not match the expected content"
+
+    # Is this some / vs \ thing?
+    # # Expected output file path
+    # expected_file_path = directory / output_file_name
+    #
+    # assert expected_file_path.exists()
+    # assert expected_file_path.is_file()
+    # # assert Path(result) == expected_file_path # varies by os!
+    # # Assertion: Check if the file is created at the correct location
+    # assert os.path.isfile(expected_file_path), "File was not created at the expected location"
+    #
+    # # Assertion: Check if the content of the file is as expected
+    # with open(expected_file_path, encoding="utf-8") as file:
+    #     content = file.read()
+    # assert content == about_content, "Content of the file does not match the expected content"
+
+
+# E       AssertionError: assert
+#   WindowsPath('C:/Users/matth/AppData/Local/Temp/pytest_of_matth/pytest_2031/test_write_to_file0/test_package/__about__.py') ==
+#   WindowsPath('C:/Users/matth/AppData/Local/Temp/pytest-of-matth/pytest-2031/test_write_to_file0/test_package/__about__.py')
+# E WindowsPath('C:/Users/matth/AppData/Local/Temp/pytest_of_matth/pytest_2031/test_write_to_file0/test_package/__about__.py')
+# =       Path('C:\\Users\\matth\\AppData\\Local\\Temp\\pytest_of_matth\\pytest_2031\\test_write_to_file0\\test_package\\__about__.py')