+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.4] - 2024-06-14
+
+### Fixed
+
+- Mypy complaints when lists are empty. Added more null safety checks.
+
+### Added
+
+- Experimental setup.py support.
+
 ## [0.1.3] - 2024-06-14
 
 ### Fixed